refactor(node): extract canvas takeover routes into src/canvas-takeover.ts (task-1773689755389)
- New: src/canvas-takeover.ts — 3 routes as Fastify plugin:
  POST /canvas/takeover, POST /canvas/takeover/release, GET /canvas/takeover
- Mutable takeover state encapsulated in module (currentTakeover)
- eventBus + queueCanvasPushEvent injected via deps
- Route-docs checker updated to scan canvas-takeover.ts
- server.ts reduced by ~137 lines
- Route-docs: 571/571, tests: 2455 pass, 0 regressions
